refacotr : 교수 타입과 매핑 분리
